// src/model/species/traits/biological/types.ts
// Positive traits
type PositiveTrait =
  | 'adaptive'
  | 'extremely_adaptive'
  | 'agrarian'
  | 'charismatic'
  | 'communal'
  | 'conformists'
  | 'conservationist'
  | 'docile'
  | 'enduring'
  | 'venerable'
  | 'industrious'
  | 'ingenious'
  | 'intelligent'
  | 'natural_engineers'
  | 'natural_physicists'
  | 'natural_sociologists'
  | 'nomadic'
  | 'quick_learners'
  | 'rapid_breeders'
  | 'resilient'
  | 'strong'
  | 'very_strong'
  | 'talented'
  | 'thrifty'
  | 'traditional'
  | 'existential_iteroparity'
  | 'incubators'
  | 'noxious'
  | 'inorganic_breath'
  | 'vocational_genomics'

// Negative traits
type NegativeTrait =
  | 'decadent'
  | 'deviants'
  | 'fleeting'
  | 'nonadaptive'
  | 'sedentary'
  | 'quarrelsome'
  | 'repugnant'
  | 'slow_breeders'
  | 'slow_learners'
  | 'solitary'
  | 'unruly'
  | 'wasteful'
  | 'weak'
  | 'jinxed'
  | 'psychological_infertility'

// Basic traits
type BasicTrait = 'hive_minded' | 'lithoid' | 'aquatic'

// Phenotype traits
type PhenotypeTrait =
  | 'phototrophic'
  | 'radiotrophic'
  | 'invasive_species'
  | 'budding'
  | 'crystallization'
  | 'gaseous_byproducts'
  | 'scintillating_skin'
  | 'volatile_excretions'
  | 'familial'
  | 'genetic_memory'
  | 'camouflage'
  | 'chromalogs'
  | 'egg_laying'
  | 'spare_organs'
  | 'seasonal_dormancy'
  | 'flight'
  | 'spatial_mastery'
  | 'shelled'
  | 'acidic_vascularity'
  | 'nascent_stage'
  | 'permeable_skin'
  | 'hollow_bones'
  | 'rooted'
  | 'brittle'

export type BiologicalTraits = PositiveTrait | NegativeTrait | BasicTrait | PhenotypeTrait